feat: AdminWrapper and components
