+++ layout/index.ts
@@ -1,2 +1,2 @@
 export { default as Warpper } from './warpper';
-export { default as AdminWarpper } from './warpper';
+export { default as AdminWarpper } from './adminWrapper';
