--- pages/admin/dashboard/file.tsx
@@ -1,14 +1,0 @@
-import React, { FC } from "react";
-
-const File: FC = () => {
-  return (
-    <div className="columns-3">
-      <div className="w-full h-52">
-        <img className="w-full" src="../../../public/company.jpg" alt="" />
-        sss
-      </div>
-    </div>
-  );
-};
-
-export default File;
