+++ pages/admin/index.tsx
@@ -1,0 +1,13 @@
+import { NextPage } from "next";
+import { AdminWarpper } from "../../layout";
+import TableMessage from "../../components/tableMessage";
+
+const Post: NextPage = () => {
+  return (
+    <AdminWarpper>
+      <TableMessage/>
+    </AdminWarpper>
+  );
+};
+
+export default Post;
